Extract the UI richText parsing for Label to be used by BooleanCollector
